compartmentalized some person components
